+++ src/BVS_probit.cpp
@@ -415,7 +415,7 @@ void BVS_probit::sampleGammaProposalRatio(
     // arma::vec proposedZ = zbinprobit(proposedBeta, dataclass);
 
     double logPriorBetaRatio = BVS_subfunc::logPDFNormal(proposedBeta, tauSq) - BVS_subfunc::logPDFNormal(betas, tauSq);
-    double logProposalBetaRatio = logP_beta - proposedBetaPrior;
+    double logProposalBetaRatio = proposedBetaPrior - logP_beta;
 
 
     // compute logLikelihoodRatio, i.e. proposedLikelihood - loglik